Remove slf4j
