+++ pg-index-health-logger/build.gradle.kts
@@ -12,7 +12,6 @@ dependencies {
 
     testImplementation(testFixtures(project(":pg-index-health-core")))
     testImplementation("org.junit.jupiter:junit-jupiter-params")
-    testImplementation(libs.logback.classic)
     testImplementation("org.mockito:mockito-core")
     testImplementation(libs.postgresql)
